chore: patch torch all the time -
"list"tests do load the modules as well - so they
need the same patch.
