+++ tests/freqai/conftest.py
@@ -1,4 +1,3 @@
-import platform
 import sys
 from copy import deepcopy
 from pathlib import Path
@@ -18,30 +17,6 @@ from tests.conftest import get_patched_exchange
 
 def is_py12() -> bool:
     return sys.version_info >= (3, 12)
-
-
-def is_mac() -> bool:
-    machine = platform.system()
-    return "Darwin" in machine
-
-
-def is_arm() -> bool:
-    machine = platform.machine()
-    return "arm" in machine or "aarch64" in machine
-
-
-@pytest.fixture(autouse=True)
-def patch_torch_initlogs(mocker) -> None:
-    if is_mac():
-        # Mock torch import completely
-        import sys
-        import types
-
-        module_name = "torch"
-        mocked_module = types.ModuleType(module_name)
-        sys.modules[module_name] = mocked_module
-    else:
-        mocker.patch("torch._logging._init_logs")
 
 
 @pytest.fixture(scope="function")
